cookbook/pocketflow-tool-search/requirements.txt
pocketflow>=0.1.0
google-search-results>=2.4.2  # SerpAPI client
openai>=1.0.0  # for search result analysis
pyyaml>=6.0.1  # for structured output
